Changes for Py2>Py3 and setup.py requires emptied
